+++ app/src/main/java/pl/futurecollars/invoicing/service/InvoiceService.java
@@ -3,16 +3,17 @@ package pl.futurecollars.invoicing.service;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 import java.util.UUID;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import pl.futurecollars.invoicing.db.Database;
-import pl.futurecollars.invoicing.db.invoices.InvoiceRepository;
+import pl.futurecollars.invoicing.repository.InvoiceRepository;
 import pl.futurecollars.invoicing.dto.InvoiceDto;
 import pl.futurecollars.invoicing.dto.mappers.InvoiceMapper;
 import pl.futurecollars.invoicing.exceptions.ConstraintException;
+import pl.futurecollars.invoicing.helpers.FilterParameters;
 import pl.futurecollars.invoicing.model.Invoice;
 import pl.futurecollars.invoicing.model.InvoiceEntry;
 
@@ -34,6 +35,9 @@ public class InvoiceService {
     }
 
     public InvoiceDto getById(UUID id) throws NoSuchElementException {
+        if(!invoiceRepository.existsById(id)) {
+            throw new NoSuchElementException("Invoice doesn't exist");
+        }
         return invoiceMapper.toDto(invoiceRepository.getById(id));
     }
 
@@ -43,20 +47,45 @@ public class InvoiceService {
                 .collect(Collectors.toList());
     }
 
-    public List<InvoiceDto> filter(Predicate<Invoice> predicate) {
+    public List<InvoiceDto> filter(FilterParameters filterParameters) {
         return invoiceRepository.findAll()
                 .stream()
-                .filter(predicate)
+                .filter(createPredicate(filterParameters))
                 .map(invoiceMapper::toDto)
                 .collect(Collectors.toList());
     }
 
     public InvoiceDto updateInvoice(InvoiceDto updatedInvoice) throws NoSuchElementException {
+        if(!invoiceRepository.existsById(updatedInvoice.getId())) {
+            throw new NoSuchElementException("Invoice doesn't exist");
+        }
         Invoice returnedInvoice = invoiceRepository.save(invoiceMapper.toEntity(updatedInvoice));
         return invoiceMapper.toDto(returnedInvoice);
     }
 
     public void deleteInvoice(UUID id) throws NoSuchElementException {
+        if(!invoiceRepository.existsById(id)) {
+            throw new NoSuchElementException("Invoice doesn't exist");
+        }
         invoiceRepository.deleteById(id);
+    }
+
+    private Predicate<Invoice> createPredicate(FilterParameters filterParameters) {
+        Predicate<Invoice> predicate = Objects::nonNull;
+        if(filterParameters.getAfterDate().isPresent()) {
+            predicate = predicate.and(invoice -> invoice.getDate().isAfter(filterParameters.getAfterDate().get()));
+        }
+        if(filterParameters.getBeforeDate().isPresent()) {
+            predicate = predicate.and(invoice -> invoice.getDate().isBefore(filterParameters.getBeforeDate().get()));
+        }
+        if(filterParameters.getSellerTaxId().isPresent()) {
+            predicate = predicate.and(invoice -> invoice.getSeller().getTaxIdentificationNumber()
+                    .equals(filterParameters.getSellerTaxId().get()));
+        }
+        if(filterParameters.getBuyerTaxId().isPresent()) {
+            predicate = predicate.and(invoice -> invoice.getBuyer().getTaxIdentificationNumber()
+                    .equals(filterParameters.getBuyerTaxId().get()));
+        }
+        return predicate;
     }
 }
